requirements.txt
#
# Don't run freeze!
# Edit by hand
#

boto==2.39.0
cryptography==2.8
Django==1.8.13
django-nose==1.4.3
django-grappelli==2.7.3
django-guardian==1.4.1
django-heroku-memcacheify==1.0.0
django-mathfilters==0.3.0
django-userena==1.4.1
django-extensions==1.6.7
dj-database-url==0.4.0
dj-static==0.0.6
djangorestframework==3.3.3
djangorestframework-xml==1.3.0
easy-thumbnails==2.3
google-api-python-client==1.4.2
gunicorn==19.4.5
humanize==0.5.1
#httplib2==0.9.2
httplib2==0.17.0
pyOpenSSL==19.1.0
inlinestyler==0.2.3
mailchimp==2.0.9
newrelic==2.60.0.46
oauth2client==1.5.2
pika==0.10.0
Pillow==2.7.0
python-dateutil==2.4.2
pycrypto==2.6.1
pygeoip==0.3.2
psycopg2==2.7.5
pylibmc==1.5.0
raven==5.10.2
requests==2.9.1
sendgrid==1.6.20
slacker==0.9.0
stripe==1.29.1
